Unify univariate and multivariate TPE
