+++ tests/samplers_tests/tpe_tests/test_parzen_estimator.py
@@ -1,122 +1,181 @@
-import itertools
-from typing import Dict
-from typing import List
+from unittest.mock import patch
 
 import numpy as np
 import pytest
 
+from optuna import distributions
 from optuna.samplers._tpe.parzen_estimator import _ParzenEstimator
-from optuna.samplers._tpe.sampler import default_weights
+from optuna.samplers._tpe.parzen_estimator import _ParzenEstimatorParameters
 
 
-class TestParzenEstimator(object):
-    @staticmethod
-    @pytest.mark.parametrize(
-        "mus, prior, magic_clip, endpoints",
-        itertools.product(
-            ([], [0.4], [-0.4, 0.4]),  # mus
-            (True, False),  # prior
-            (True, False),  # magic_clip
-            (True, False),  # endpoints
-        ),
+SEARCH_SPACE = {
+    "a": distributions.UniformDistribution(1.0, 100.0),
+    "b": distributions.LogUniformDistribution(1.0, 100.0),
+    "c": distributions.DiscreteUniformDistribution(1.0, 100.0, 3.0),
+    "d": distributions.IntUniformDistribution(1, 100),
+    "e": distributions.IntLogUniformDistribution(1, 100),
+    "f": distributions.CategoricalDistribution(["x", "y", "z"]),
+}
+
+MULTIVARIATE_SAMPLES = {
+    "a": np.array([1.0]),
+    "b": np.array([1.0]),
+    "c": np.array([1.0]),
+    "d": np.array([1]),
+    "e": np.array([1]),
+    "f": np.array([1]),
+}
+
+_PRECOMPUTE_SIGMAS0 = "optuna.samplers._tpe.parzen_estimator._ParzenEstimator._precompute_sigmas0"
+
+
+@pytest.mark.parametrize("consider_prior", [True, False])
+def test_init_parzen_estimator(consider_prior: bool) -> None:
+
+    parameters = _ParzenEstimatorParameters(
+        consider_prior=consider_prior,
+        prior_weight=1.0,
+        consider_magic_clip=False,
+        consider_endpoints=False,
+        weights=lambda x: np.arange(x) + 1.0,
     )
-    def test_calculate_shape_check(
-        mus: List[float], prior: bool, magic_clip: bool, endpoints: bool
-    ) -> None:
 
-        s_weights, s_mus, s_sigmas = _ParzenEstimator._calculate(
-            np.array(mus),
-            -1.0,
-            1.0,
-            prior_weight=1.0,
-            consider_prior=prior,
-            consider_magic_clip=magic_clip,
-            consider_endpoints=endpoints,
-            weights_func=default_weights,
+    with patch(_PRECOMPUTE_SIGMAS0, return_value=np.ones(1)):
+        mpe = _ParzenEstimator(MULTIVARIATE_SAMPLES, SEARCH_SPACE, parameters)
+
+    weights = np.array([1] + consider_prior * [1], dtype=float)
+    weights /= weights.sum()
+    q = {"a": None, "b": None, "c": 3.0, "d": 1.0, "e": None, "f": None}
+    low = {"a": 1.0, "b": np.log(1.0), "c": -0.5, "d": 0.5, "e": np.log(0.5), "f": None}
+    high = {"a": 100.0, "b": np.log(100.0), "c": 101.5, "d": 100.5, "e": np.log(100.5), "f": None}
+
+    assert np.all(mpe._weights == weights)
+    assert mpe._q == q
+    assert mpe._low == low
+    assert mpe._high == high
+
+    expected_sigmas = {
+        "a": [99.0] + consider_prior * [99.0],
+        "b": [np.log(100.0)] + consider_prior * [np.log(100)],
+        "c": [102.0] + consider_prior * [102.0],
+        "d": [100.0] + consider_prior * [100.0],
+        "e": [np.log(100.5) - np.log(0.5)] + consider_prior * [np.log(100.5) - np.log(0.5)],
+        "f": None,
+    }
+    expected_mus = {
+        "a": [1.0] + consider_prior * [50.5],
+        "b": [np.log(1.0)] + consider_prior * [np.log(100) / 2.0],
+        "c": [1.0] + consider_prior * [50.5],
+        "d": [1.0] + consider_prior * [50.5],
+        "e": [np.log(1.0)] + consider_prior * [(np.log(100.5) + np.log(0.5)) / 2.0],
+        "f": None,
+    }
+    expected_categorical_weights = {
+        "a": None,
+        "b": None,
+        "c": None,
+        "d": None,
+        "e": None,
+        "f": np.array([[0.2, 0.6, 0.2], [1.0 / 3.0, 1.0 / 3.0, 1.0 / 3.0]])
+        if consider_prior
+        else np.array([[0.25, 0.5, 0.25]]),
+    }
+
+    for param_name in mpe._sigmas:
+        np.testing.assert_equal(
+            mpe._sigmas[param_name],
+            expected_sigmas[param_name],
+            err_msg='parameter "{}"'.format(param_name),
+        )
+        np.testing.assert_equal(
+            mpe._mus[param_name],
+            expected_mus[param_name],
+            err_msg="parameter: {}".format(param_name),
+        )
+        np.testing.assert_equal(
+            mpe._categorical_weights[param_name],
+            expected_categorical_weights[param_name],
+            err_msg="parameter: {}".format(param_name),
         )
 
-        # Result contains an additional value for a prior distribution if prior is True or
-        # len(mus) == 0 (in this case, prior is always used).
-        assert len(s_weights) == len(mus) + int(prior) if len(mus) > 0 else len(mus) + 1
-        assert len(s_mus) == len(mus) + int(prior) if len(mus) > 0 else len(mus) + 1
-        assert len(s_sigmas) == len(mus) + int(prior) if len(mus) > 0 else len(mus) + 1
 
-    # TODO(ytsmiling): Improve test coverage for weights_func.
-    @staticmethod
-    @pytest.mark.parametrize(
-        "mus, flags, expected",
-        [
-            [
-                [],
-                {"prior": False, "magic_clip": False, "endpoints": True},
-                {"weights": [1.0], "mus": [0.0], "sigmas": [2.0]},
-            ],
-            [
-                [],
-                {"prior": True, "magic_clip": False, "endpoints": True},
-                {"weights": [1.0], "mus": [0.0], "sigmas": [2.0]},
-            ],
-            [
-                [0.4],
-                {"prior": True, "magic_clip": False, "endpoints": True},
-                {"weights": [0.5, 0.5], "mus": [0.0, 0.4], "sigmas": [2.0, 0.6]},
-            ],
-            [
-                [-0.4],
-                {"prior": True, "magic_clip": False, "endpoints": True},
-                {"weights": [0.5, 0.5], "mus": [-0.4, 0.0], "sigmas": [0.6, 2.0]},
-            ],
-            [
-                [-0.4, 0.4],
-                {"prior": True, "magic_clip": False, "endpoints": True},
-                {"weights": [1.0 / 3] * 3, "mus": [-0.4, 0.0, 0.4], "sigmas": [0.6, 2.0, 0.6]},
-            ],
-            [
-                [-0.4, 0.4],
-                {"prior": True, "magic_clip": False, "endpoints": False},
-                {"weights": [1.0 / 3] * 3, "mus": [-0.4, 0.0, 0.4], "sigmas": [0.4, 2.0, 0.4]},
-            ],
-            [
-                [-0.4, 0.4],
-                {"prior": False, "magic_clip": False, "endpoints": True},
-                {"weights": [0.5, 0.5], "mus": [-0.4, 0.4], "sigmas": [0.8, 0.8]},
-            ],
-            [
-                [-0.4, 0.4, 0.41, 0.42],
-                {"prior": False, "magic_clip": False, "endpoints": True},
-                {
-                    "weights": [0.25, 0.25, 0.25, 0.25],
-                    "mus": [-0.4, 0.4, 0.41, 0.42],
-                    "sigmas": [0.8, 0.8, 0.01, 0.58],
-                },
-            ],
-            [
-                [-0.4, 0.4, 0.41, 0.42],
-                {"prior": False, "magic_clip": True, "endpoints": True},
-                {
-                    "weights": [0.25, 0.25, 0.25, 0.25],
-                    "mus": [-0.4, 0.4, 0.41, 0.42],
-                    "sigmas": [0.8, 0.8, 0.4, 0.58],
-                },
-            ],
-        ],
+def test_sample_parzen_estimator() -> None:
+
+    parameters = _ParzenEstimatorParameters(
+        consider_prior=False,
+        prior_weight=0.0,
+        consider_magic_clip=False,
+        consider_endpoints=False,
+        weights=lambda x: np.arange(x) + 1.0,
     )
-    def test_calculate(
-        mus: List[float], flags: Dict[str, bool], expected: Dict[str, List[float]]
-    ) -> None:
+    with patch(_PRECOMPUTE_SIGMAS0, return_value=1e-8 * np.ones(2)):
+        mpe = _ParzenEstimator(MULTIVARIATE_SAMPLES, SEARCH_SPACE, parameters)
 
-        s_weights, s_mus, s_sigmas = _ParzenEstimator._calculate(
-            np.array(mus),
-            -1.0,
-            1.0,
-            prior_weight=1.0,
-            consider_prior=flags["prior"],
-            consider_magic_clip=flags["magic_clip"],
-            consider_endpoints=flags["endpoints"],
-            weights_func=default_weights,
-        )
+    # Test the shape of the samples.
+    output_samples = mpe.sample(np.random.RandomState(0), 3)
+    for param_name in output_samples:
+        assert output_samples[param_name].shape == (3,)
 
-        # Result contains an additional value for a prior distribution if consider_prior is True.
-        np.testing.assert_almost_equal(s_weights, expected["weights"])
-        np.testing.assert_almost_equal(s_mus, expected["mus"])
-        np.testing.assert_almost_equal(s_sigmas, expected["sigmas"])
+    # Test the values of the output.
+    # As we set ``consider_prior`` = False and pre-computed sigma to be 1e-8,
+    # the samples almost equals to the input ``MULTIVARIATE_SAMPLES``.
+    output_samples = mpe.sample(np.random.RandomState(0), 1)
+    for param_name, samples in output_samples.items():
+        if samples.dtype == str:
+            assert samples[0] == "y", "parameter {}".format(param_name)
+        else:
+            np.testing.assert_almost_equal(
+                samples,
+                MULTIVARIATE_SAMPLES[param_name],
+                decimal=2,
+                err_msg="parameter {}".format(param_name),
+            )
+
+    # Test the output when the seeds are fixed.
+    assert output_samples == mpe.sample(np.random.RandomState(0), 1)
+
+
+def test_suggest_with_step_parzen_estimator() -> None:
+    parameters = _ParzenEstimatorParameters(
+        consider_prior=False,
+        prior_weight=0.0,
+        consider_magic_clip=False,
+        consider_endpoints=False,
+        weights=lambda x: np.arange(x) + 1.0,
+    )
+
+    # Define search space for distribution with step argument and true ranges
+    search_space = {
+        "c": distributions.DiscreteUniformDistribution(low=1.0, high=7.0, q=3.0),
+        "d": distributions.IntUniformDistribution(low=1, high=5, step=2),
+    }
+    multivariate_samples = {"c": np.array([4]), "d": np.array([1])}
+    valid_ranges = {"c": set(np.arange(1.0, 10.0, 3.0)), "d": set(np.arange(1, 7, 2))}
+
+    with patch(_PRECOMPUTE_SIGMAS0, return_value=np.ones(2)):
+        mpe = _ParzenEstimator(multivariate_samples, search_space, parameters)
+
+    # Draw 10 samples, and check if all valid values are sampled.
+    output_samples = mpe.sample(np.random.RandomState(0), 10)
+    for param_name in output_samples:
+        assert set(output_samples[param_name]) == valid_ranges[param_name]
+
+
+def test_log_pdf_parzen_estimator() -> None:
+    parameters = _ParzenEstimatorParameters(
+        consider_prior=False,
+        prior_weight=1.0,
+        consider_magic_clip=True,
+        consider_endpoints=True,
+        weights=lambda x: np.arange(x) + 1.0,
+    )
+    # Parzen estimator almost becomes mixture of Dirac measures.
+    with patch(_PRECOMPUTE_SIGMAS0, return_value=1e-8 * np.ones(1)):
+        mpe = _ParzenEstimator(MULTIVARIATE_SAMPLES, SEARCH_SPACE, parameters)
+
+    log_pdf = mpe.log_pdf(MULTIVARIATE_SAMPLES)
+    output_samples = mpe.sample(np.random.RandomState(0), 100)
+    output_log_pdf = mpe.log_pdf(output_samples)
+    # The likelihood of the previous observations is a positive value, and that of the points
+    # sampled by the Parzen estimator is almost zero.
+    assert np.all(log_pdf >= output_log_pdf)
